HRV.jl

HRV.infile — Function

This function reads a file (txt or csv) and saves the data in an array :param file: is the location of the input file :return: an array containing the NN intervals

This function reads a wfdb file and saves the data in an array :param record: is the name of the record :param annotator: is the annotator of the record :return: an array containing the NN intervals

HRV.time_domain — Function

This function calculates the time domain analysis methods and prints the results :param n: is the array that contains the NN-inetrvals

HRV.frequency — Function

This function calculates the frequency analysis methods and prints the results :param n: is the array that contains the NN-inetrvals

HRV.nonlinear — Function

This function calculates the nonlinear analysis methods and prints the results

Arguments:

n : is the array that contains the NN-inetrvals

m : is the embedding dimension, default=2

r : is the tolerance, default=6